changed how I compute bf data
before, I just passed a process number to the same binary and let it compute that split of the data. Parallelization was achived through the launch of multiple programs with different positional parameters. I changed the script so that is uses Rusts concurrency features. It was surprisingly easy to set-up. Now I only need to pass the number of threads I want to spawn, and the rest is done by the binary.

+++ hnsw/src/helpers/data.rs
@@ -3,10 +3,10 @@ use std::fs;
 use std::fs::File;
 use std::io::{BufReader, Result};
 
-pub fn split_vector(vector: Vec<i32>, nb_splits: usize, split_to_compute: usize) -> Vec<i32> {
+pub fn split_vector(vector: Vec<i32>, nb_splits: u8, split_to_compute: u8) -> Vec<i32> {
     let mut split_vector: Vec<Vec<i32>> = Vec::new();
 
-    let per_split = vector.len() / nb_splits;
+    let per_split = vector.len() / nb_splits as usize;
 
     let mut buffer = 0;
     for idx in (0..nb_splits).into_iter() {
@@ -25,7 +25,7 @@ pub fn split_vector(vector: Vec<i32>, nb_splits: usize, split_to_compute: usize)
 
     assert!(sum_lens == vector.len(), "sum: {sum_lens}");
 
-    split_vector[split_to_compute].to_owned()
+    split_vector[split_to_compute as usize].to_owned()
 }
 
 pub fn load_bf_data(dim: usize, lim: usize) -> Result<HashMap<usize, Vec<(usize, f32)>>> {
